Create day8.py
Affirmations (or insults) Generator
Today's focus is using all the skills you have learned so far:

input and output
concatenation
if statements
nested if statements

Build a custom affirmations generator to give the user a customized affirmation each day of the week.

Make sure you ask the user for their name, the current day of the week, and a few of their favourite things. All of this information should be used to help you build your affirmations.
The goal is to generate a unique message for each day of the week based on the user's answers.
Use concatenation to generate the affirmation.

If affirmations are not your jam, then you could do a daily joke or insult. Please don't be mean. Keep it light and funny.

Final challenge: Can you create if statements that do not care about capital or lowercase letters of names?